move functions related to streaming to inference_stream
